feat: FileBrowser refactoring and test improvements
- Created unified pkg/filebrowser package for code reuse
- Eliminated 190+ lines of duplicate FileBrowser code (29% reduction)
- Refactored UserStorageProxy to use shared proxy handler
- Updated FileBrowserManager and UserStorageManager
- Fixed K8s integration test pod deletion timing issue
- Added WaitForPodDeleted helper method
- All tests passing (unit + integration)
- Documentation: FILEBROWSER_REFACTORING_SUMMARY.md
- Migration guide: FILEBROWSER_MIGRATION_GUIDE.md

+++ internal/api/handlers/group_handler.go
@@ -54,7 +54,7 @@ func (h *GroupHandler) GetGroupByID(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, response.ErrorResponse{Error: "invalid group id"})
 		return
 	}
-	group, err := h.svc.GetGroup(uint(gid))
+	group, err := h.svc.GetGroup(gid)
 	if err != nil {
 		c.JSON(http.StatusNotFound, response.ErrorResponse{Error: "group not found"})
 		return
